Provide mechanism to shape routing behavior via policies/profiles
Problem: Policies and profiles can be configured to shape the virtual
server routing traffic and to enable/disable functionality relating to
gorouter functionality.

Solution: Provide configuration areas to attach pre-configured policy
and profile to routing virtuals. The controller creates the routing
policy and therefore allows for a prerouting policy list and a
postrouting policy list.

The controller adds no profiles and therefore this is a single list
which is passed through to the bigip.