Update route policy on endpoint prune
Problem: A registry prune would not cause updates to the
routing policy.

Solution: Call update functions to either remove the rule when
the pool is empty or to update pool when endpoints have been
pruned.

Re-worked logic to not save all pools and endpoints in the f5router
as they are already information in the Trie. Only create and save
policy rules. The Trie is converted into pool configurations when
configuration is written to disk.