Fix intermittent main_test test failures
Problem: Some main_test tests fail intermittently.

Solution: golang exec / wait will sometimes report a process
as closed with the returned exit code or sometimes report a
process as signaled with the signum when the ctlr signals
the subprocess to close.

If the wait reports the process is signaled and the ctlr
caused the signal then execute normally. Only consider a signal
as extraordinary if the ctlr hadn't signaled the subprocess.

Also, don't use kill to close the process, this leaves subprocesses
around. Use SIGTERM.