Route unregisters should emit Update with remaining pool members
Problem: When route unregisters are received the registry unregister
call only calls Remove. Remove implies the route is dead and not
simply updated. This causes pool members to fail to be updated
because the Remove pathway expects empty pools or deleted URIs.

Solution: Call Update when there are still endpoints available
after a route unregister. Only when all endpoints are unregistered
should the Remove case be called.

Unit tests ran