Deleting or unmapping routes may not remove pools
Problem: When a route is unmapped from an application removing all
endpoints or when an app is deleted; also removing all endpoints.
The corresponding BigIP pool may not be deleted along with the
policy rule.

Solution: When the registry pools are empty they should be deleted
from the registry. This happened after the route listener update.
The queue processor may generate a new config before the original
thread deleted the uri entry.

Move the uri delete before the update. Ensure all state is updated
before notifying update.