Add support for TCP routing config on the BIG-IP device
Problem:
Controller was listening to TCP events but was not including TCP objects
in the config
f5router accessing storage objects directly
Controller needs to keep a state of the config and update this object

Solution:
Add a cached version to f5router of the current state of the world then
update this version with each update. This cuts down on the process time
for each update and decouples f5router from the data structures
currently in use. Add TCP routing objects to config being output and add
ability to create these objects in f5router including pools and
virtuals.