Add vip2vip configuration (#56)
Problem:
The controller currently only has two vips for all http traffic. This is
restrictive as all config applies to all routes.

Solution:
The is the first step in allowing a per route config for cf.
This implements a vip targeting vip setup with a tier 1 routing vip
passing requests to the tier 2 vip.
Added additional tests and updated current tests.