Fix version information in documentation
Problem:
The docs version string as displayed on the clouddocs sidebar for
product docs shows an incorrect version string.

Solution:
The problem is that the docs version string is based on the version
information from next-version.txt. This is okay for non-release branches
such as master but not accurate for the stable branches off of which we
pull releases.

The fix is to use the latest git tag to determine docs version for
stable branches.