Added success logging flag.
Problem:
The retry callback cbs were logging success the first time that they
succeeded. This is incorrect since the first time the cb is called a
retry has not been made so the code executed without error so there is
not a need to log a successful outcome.

Solution:
Add a success logging flag that will only log success if the cb has
been called more than once in the backoff loop (which means that the
cb had failed at least once necessitating the success logging).